+++ tests/conftest.py
@@ -59,6 +59,7 @@ def connection_factory() -> Generator[ConnectionFactory, None, None]:
         database: str | None = None,
         result_page_fetch_pause_millis: int = 100,
         http_user_agent: str | None = None,
+        endpoint: str | None = None,
     ) -> Connection:
         if flink_api_key is None:
             flink_api_key = os.getenv("CONFLUENT_FLINK_API_KEY", "")
@@ -88,6 +89,7 @@ def connection_factory() -> Generator[ConnectionFactory, None, None]:
             database=database,
             result_page_fetch_pause_millis=result_page_fetch_pause_millis,
             http_user_agent=http_user_agent,
+            endpoint=endpoint,
         )
 
         connections.append(connection)